% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_wrappers.R
\name{bq_get}
\alias{bq_get}
\alias{bq_post}
\alias{bq_post_steps}
\title{bq_get}
\usage{
bq_get(query, project = NULL, location = NULL, ...)

bq_post(
  data,
  project,
  dataset,
  table,
  create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_TRUNCATE",
  quiet = F
)

bq_post_steps(data, steps_by = 1e+05, ...)
}
\arguments{
\item{query}{SQL query to execute}

\item{project, dataset, table}{str name of the entity that will perform the job}

\item{location}{str name of the wkt column to convert to sf.}

\item{data}{df to upload.}

\item{create_disposition, write_disposition, quiet}{options to set in the job}

\item{steps_by}{number of rows to upload in each step.}
}
\value{
invisible.
}
\description{
wrapper helpers to communicate with Bigquery
}
