% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_get_urban_vars.R
\name{bq_get_population}
\alias{bq_get_population}
\alias{bq_get_competitors}
\alias{bq_get_wawa_movements}
\alias{bq_get_homicidios}
\alias{bq_get_property_value}
\title{bq_get_urban_vars}
\usage{
bq_get_population(coord, buffer = NULL, get_hex = TRUE)

bq_get_competitors(coord, buffer = NULL)

bq_get_wawa_movements(coord, buffer = NULL)

bq_get_homicidios(coord, buffer = NULL)

bq_get_property_value(coord, buffer = NULL)
}
\arguments{
\item{coord}{sf point object wit lat-lon coordinate in WGS84 CRS}

\item{buffer}{numeric buffer in meters to apply to the coordinate}

\item{get_hex}{logical if TRUE return the values by hexagons, otherwise aggregates values to a single joined hexagons}
}
\value{
invisible.
}
\description{
Perform queries against BQ tables to get urban variables for a selected coordinate
}
