% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geocode_address}
\alias{geocode_address}
\alias{geocode_apply}
\alias{geocode_google_safe}
\alias{geocode_update}
\title{geocoding}
\usage{
geocode_address(address_df, address_column)

geocode_apply(address_df)

geocode_google_safe(address_str)

geocode_update(address_df, id, lat, lng, place_id)
}
\arguments{
\item{address_df}{dataframe with the column string *address*}

\item{address_column}{name of the column containning addresses}

\item{address_str}{string Address to geocode}

\item{id, lat, lng, place_id}{values to update in *address_df* by id}
}
\value{
geocode_address, geocode_apply and geocode_update return the same address_df with lat, lng and place_id columns
geocode_google_safe return a list with lat, lng and place_id
}
\description{
apply geocode to an address text via google Geocode API
}
