% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_google_places.R
\name{get_google_places}
\alias{get_google_places}
\alias{get_google_places_pagination}
\alias{bind_places}
\title{get_google_places}
\usage{
get_google_places(...)

get_google_places_pagination(...)

bind_places(places_list)
}
\arguments{
\item{...}{parameters to pass to googleway::get_places}

\item{places_list}{list of places returned by calling `get_google_places` in a map routine}
}
\value{
`get_google_places` returns tibble. `get_google_places_pagination` returns response list. `bind_places` returns a tibble
}
\description{
Wrapper of googleway::get_places to paginate and get all possible results.
}
\details{
`get_google_places` cleans the response to get a tibble of places and attributes:
name, place_id, types, vicinity, rating, user_ratings_total, lat, lng
`get_google_places_pagination` returns a list with max 3 responses (Google Places API return max 3 resposes of 20 places each)
`bind_places` is a cleaning wrapper to standarize responses of `get_google_places` into a tibble
}
