% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_plot.R
\name{leaflet_plot}
\alias{leaflet_plot}
\alias{map_polygons}
\title{geocoding}
\usage{
leaflet_plot(data_sf)

map_polygons(
  table_sf,
  variable_color,
  variable_label,
  legend_title = NULL,
  legend_prefix = NULL,
  legend_suffix = NULL,
  color_palette = "YlOrRd"
)
}
\arguments{
\item{data_sf}{dataframe with the column string *address*}
}
\value{
geocode_address, geocode_apply and geocode_update return the same address_df with lat, lng and place_id columns
geocode_google_safe return a list with lat, lng and place_id
}
\description{
quick leaflet map with feature type discovery
}
