% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_wrappers.R
\name{sf_df_to_list}
\alias{sf_df_to_list}
\alias{sf_to_wk}
\title{sf_df_to_list}
\usage{
sf_df_to_list(data_sf, names_from = NULL, crs = 4326)

sf_to_wk(data_sf, location = geometry)
}
\arguments{
\item{data_sf}{sf dataframe}

\item{names_from}{column of the field that will be passing to list names. Names must be uniques}

\item{crs}{crs to apply to the new list elements}

\item{location}{geometry field to be converted to wkt}
}
\value{
invisible.
}
\description{
wrapper helpers to apply sf functions
}
