% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vktools-package.R
\docType{package}
\name{vktools-package}
\alias{vktools}
\alias{vktools-package}
\title{vktools: Wrappers and utils for Vikua Corp analysis}
\description{
Vikua is a urban-tech latam startup focused on provide urban-oriented analytic solutions This package groups common functions for data analysis in Vikua Corp.
}
\author{
\strong{Maintainer}: Rodrigo Serrano \email{roserrano@vikua.com}

}
\keyword{internal}
