% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bq_wrappers.R
\name{bq_get}
\alias{bq_get}
\alias{sf_to_wk}
\alias{bq_post}
\title{bq_get}
\usage{
bq_get(query, dataset = NULL, location = NULL)

sf_to_wk(data, location = geometry)

bq_post(
  data,
  project,
  dataset,
  table,
  create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_TRUNCATE",
  quiet = F
)
}
\arguments{
\item{query}{SQL query to execute}

\item{dataset}{str name of the dataset that will perform the job}

\item{location}{str name of the wkt column to convert to sf. When bq_post the column is a geometry in a `data`_sf object and it will be converted to wkt}

\item{data}{df to upload. In `sf_to_wk` data must be a sf object}

\item{project}{str name of the project that will perform the job}

\item{table}{str name of the table that will perform the job}
}
\value{
invisible.
}
\description{
wrapper helpers to communicate with Bigquery
}
