% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_plot.R
\name{leaflet_plot}
\alias{leaflet_plot}
\alias{map_polygons}
\title{leaflet_plot}
\usage{
leaflet_plot(data_sf)

map_polygons(
  data_sf,
  variable_color,
  variable_label,
  palette = "YlOrRd",
  reverse = FALSE,
  legend_title = NULL,
  legend_prefix = NULL,
  legend_suffix = NULL,
  weight = 1,
  fillOpacity = 0.6,
  color = "#444444",
  highlightOptions = leaflet::highlightOptions(color = "white", weight = 2, bringToFront
    = TRUE)
)
}
\arguments{
\item{data_sf}{sf dataframe}

\item{variable_color, variable_label}{fields to use in map}

\item{palette, reverse}{settings in leaflet::colorNumeric}

\item{legend_title, legend_prefix, legend_suffix, weight, fillOpacity, color, highlightOptions}{settings. view ?leaflet() fro more help}
}
\value{
leaflet map viewer
}
\description{
quick leaflet map with feature type discovery
}
